Fixed wrong type Wind parameter --> deg

+++ app/src/main/java/ch/ti8m/gol/bbw_route/domain/entity/openweathermap/Wind.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
     @SerializedName("speed") val speed: Double,
-    @SerializedName("deg") val direction: Int
+    @SerializedName("deg") val direction: Double
 )
